refactor(cli): write tests for execution logs
