+++ packages/cdktf-cli/lib/cdktf-project.ts
@@ -7,6 +7,7 @@ import { TerraformPlan } from "./models/terraform";
 import { NestedTerraformOutputs } from "./output";
 import { logger } from "./logging";
 import minimatch from "minimatch";
+import { enhanceLogMessage } from "./execution-logs";
 
 type MultiStackApprovalUpdate = {
   type: "waiting for approval";
@@ -280,7 +281,7 @@ export type ExecutionOptions = MultipleStackOptions & {
 
 type LogMessage = {
   stackName: string;
-  messageWithConstructPath: string;
+  messageWithConstructPath?: string;
   message: string;
 };
 
@@ -290,69 +291,6 @@ type Buffered<T, V> = {
   value: T;
   type: V;
 };
-
-type TFIdentifier = string;
-type CDKTFResourcePath = string;
-type TfResource = Record<
-  string,
-  {
-    ["//"]?: {
-      metadata?: {
-        path?: string;
-      };
-    };
-  }
->;
-type TfResourceType = Record<string, TfResource>;
-type TFJson = {
-  data: TfResourceType;
-  resource: TfResourceType;
-};
-export function enhanceLogMessage(message: string, stack: SynthesizedStack) {
-  // we never want to throw, if it does not work we just do as if it did not happen
-  try {
-    const json = JSON.parse(stack.content) as TFJson;
-    const pathMapping: Record<TFIdentifier, CDKTFResourcePath> = {};
-
-    ["data", "resource"].forEach((type) => {
-      Object.entries(json[type as "data" | "resource"] || {}).forEach(
-        ([resourceType, resourceInstances]) => {
-          Object.entries(resourceInstances || {}).forEach(
-            ([resourceName, resource]) => {
-              // Some constructs don't have this metadata
-              if (
-                resource["//"] &&
-                resource["//"].metadata &&
-                resource["//"].metadata.path
-              ) {
-                pathMapping[`${resourceType}.${resourceName}`] = resource[
-                  "//"
-                ].metadata.path.replace(`${stack.name}/`, "");
-              }
-            }
-          );
-        }
-      );
-    });
-
-    return message
-      .split("\n")
-      .map((line) => {
-        const matchingEntry = Object.entries(pathMapping).find(
-          ([tfIdentifier]) => line.includes(tfIdentifier)
-        );
-        if (!matchingEntry) {
-          return line;
-        }
-        const [id, cdkPath] = matchingEntry;
-        return line.replace(id, `${id} (${cdkPath})`);
-      })
-      .join("\n");
-  } catch (e) {
-    logger.debug(`Could not enhance log message: ${e}`);
-    return message;
-  }
-}
 
 export class CdktfProject {
   public stacks?: SynthesizedStack[];
